[#816] Pulled pure workspace nav functions out into separate utils namesapce
